Various cleanup

+++ src/main/java/net/katsstuff/danmakucore/client/handler/SpellcardHandler.java
@@ -53,7 +53,7 @@ public class SpellcardHandler {
 			int stringWidth = mc.fontRenderer.getStringWidth(formattedText);
 
 			if(info.shouldMirrorText()) {
-				mc.fontRenderer.drawStringWithShadow(formattedText, renderX - stringWidth, renderY, 0xFFFFFF);
+				mc.fontRenderer.drawStringWithShadow(formattedText, renderX - (float)stringWidth, renderY, 0xFFFFFF);
 				Gui.drawRect(renderX - stringWidth, renderY + fontHeight, renderX - 2, renderY + fontHeight + 1, red);
 				Gui.drawRect(renderX - stringWidth + 40, renderY + 2 + fontHeight, renderX - 2, renderY + fontHeight + 3, red);
 			}
